% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{checkGeneMetaFormat}
\alias{checkGeneMetaFormat}
\title{Check GeneMeta format}
\usage{
checkGeneMetaFormat(GeneMeta)
}
\arguments{
\item{GeneMeta}{Data frame that contains the project metadata. See \link{createGeneMeta}}
}
\value{
print error messages or PASS with summary of data
}
\description{
Check to make sure the GeneMeta object is formated correctly and return summary of object
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
