% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{color.bar}
\alias{color.bar}
\title{Plot heatmap color scale}
\usage{
color.bar(lut, min, max=-min, nticks=11, title='')
}
\arguments{
\item{lut}{User specified color ramp}

\item{min}{Minimum value}

\item{max}{Maximum value}

\item{nticks}{Number of ticks}

\item{title}{Title for color scale.}
}
\description{
Plot a user defined color scale
}
\author{
Matthew Zinkgraf, \email{Matthew.Zinkgraf@wwu.edu}
}
