% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createGeneMeta}
\alias{createGeneMeta}
\title{Create gene metadata object}
\usage{
createGeneMeta(rpkm)
}
\arguments{
\item{rpkm}{List object containing the rpkm values for each species.}
}
\description{
This function generates a data frame that contains the species and gene information, and columns to store module results.
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
