% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filterCommunityAssign}
\alias{filterCommunityAssign}
\title{Filter Louvain community assignemnts}
\usage{
filterCommunityAssign(results,minMem=10)
}
\arguments{
\item{results}{Data frame containing the community assignments for each genes and contains nRuns+1 columns. The first column contains an integer referencing the GeneMeta ID}

\item{minMem}{Minimum number of genes in a community to be considered significant. Default = 10}
}
\value{
Returns a sparse matrix containing to occurance of each gene in each community. 1 = present and 0 = absent
}
\description{
This function removes communities from the sparse matrix that have a minimum number of gene members
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
