% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filterVariance}
\alias{filterVariance}
\title{Filter expression based on variance}
\usage{
filterVariance(rpkm, variance=0.1)
}
\arguments{
\item{rpkm}{List object containing the rpkm values for each species.}

\item{variance}{Variance threshold.}
}
\description{
This function filters rpkm expression data based on user defined variance threshold.
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
