% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getEdgelist}
\alias{getEdgelist}
\title{Create an edgelist from rpkm expression values}
\usage{
getEdgelist(rpkm, GeneMeta, top=5, weight=1,
nThreads = 3, parallel_apply=FALSE)
}
\arguments{
\item{rpkm}{A list object where each element in the list is a data frame of rpkm values for each species}

\item{GeneMeta}{Data frame that contains the project metadata. See \link{createGeneMeta}}

\item{top}{An integer specifying the number of neighbors for each gene that should be printed to the edgelist. Default = 5}

\item{weight}{The edge weight between genes. Default = 1}

\item{nThreads}{The number of multiple threads that should be used to calculate the correlation matrix. Default = 3}

\item{parallel_apply}{Logical value indicating if parRapply from parallel should be used in calulation. Default = FALSE}
}
\description{
Generate an edgelist for each gene from a list of rpkm values and return each gene and its top most correlated neighbors
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
