% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getOrthoWeights}
\alias{getOrthoWeights}
\title{Convert orthologous relationships to edgelist}
\usage{
getOrthoWeights(ortho, GeneMeta, couple_const=1)
}
\arguments{
\item{ortho}{List object that contains the orthologous gene relationships between species}

\item{GeneMeta}{Data frame that contains the project metadata. See \link{createGeneMeta}}

\item{couple_const}{A contant that determines the relative contribution of orthologous gene relationships between species. Default=1}
}
\description{
Generate a coupling matrix that connects species based on the orthologus relationships between species. Relationships can range from one-to-one to many-to-many
}
\references{
Koon-Kiu Yan, Daifeng Wang, Joel Rozowsky, Henry Zheng, Chao Cheng and Mark Gerstein. 2014. OrthoClust: an orthology-based network framework for clustering data across multiple species. Genome Biology. 15:R100
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
#Create ortho
data("GeneMeta")
data("potri_eugra")
data("potri_sapur")
data("sapur_eugra")

ortho <- list(potri_eugra = potri_eugra,
           potri_sapur = potri_sapur, sapur_eugra = sapur_eugra)

#calculate orthologous weights
ortho_weights <- getOrthoWeights(ortho, GeneMeta, couple_const = 1)
}
\keyword{gene}
\keyword{ortholog}
