% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiSppHclust}
\alias{multiSppHclust}
\title{Hierarchical clustering of gene co-appearance in Louvain communities}
\usage{
multiSppHclust(occurance, nRuns, GeneMeta)
}
\arguments{
\item{occurance}{Sparse matrix that contain the presence (1) and absence (0) of each gene in each Louvain community.}

\item{nRuns}{Integer specifying the number of Louvain runs.}

\item{GeneMeta}{Data frame that contains the project metadata. See \link{createGeneMeta}}
}
\description{
For each species calculate a co-appearance matrix representing how often genes are assigned to the same Louvain communities and perform hierarchical clustering to determine which genes have similar co-appearance.
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
