% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiSppModules}
\alias{multiSppModules}
\title{Identify modules with high co-appearance}
\usage{
multiSppModules(multiSpp_results, GeneMeta, minModuleSize, cut)
}
\arguments{
\item{multiSpp_results}{A list object containing the gene order and hclust dendrograms for each species. See \link{multiSppHclust}}

\item{GeneMeta}{Data frame that contains the project metadata. See \link{createGeneMeta}}

\item{minModuleSize}{A vector containing the minimum module size to be used for each species in the data set.}

\item{cut}{A vector containing the cut height to be used for species in the data set}
}
\description{
Use cutreeDynamic to identify gene modules that have high co-appearance.
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
