% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multMerge}
\alias{multMerge}
\title{Combine multiple files}
\usage{
multMerge(mypath, pattern="*\\.out", header=FALSE, sep=" ")
}
\arguments{
\item{mypath}{Path to folder.}

\item{pattern}{Unique pattern that matches files to be be read and combined using perl style regular expressions.}

\item{header}{A logical value indicateing if the file contains a header row. Default = FALSE}

\item{sep}{The field separator character.}
}
\description{
Read and combine multiple results files from a folder. The results are combined using cbind and the row order in each file must be the same.
}
\seealso{
\code{\link[base]{list.files}}
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
