% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multMergeHTseq}
\alias{multMergeHTseq}
\title{Combine multiple htseq files using merge}
\usage{
multMergeHTseq(mypath, pattern="*\\\\.htseq", byY="gene",
    RegEx="(\\\\w+)\\\\.htseq\\\\.txt", Replace="\\\\1", sep="\t")
}
\arguments{
\item{mypath}{Path to folder.}

\item{pattern}{Unique pattern that matches files to be read and merged using perl style regular expressions.}

\item{byY}{Column name that should be used in the merge function.}

\item{RegEx}{Perl style regular expression that matches pattern in the filename. Used to extract library name from filename. Example "(\\\\w+)\\\\.htseq\\\\.txt" matches Library1 in filename Library1.htseq.txt}

\item{Replace}{Perl style replacement. Default = "\\\\1"}

\item{sep}{The field separator character.}
}
\description{
Read and combine tab deliminted htseq results files from a folder using the merge function.
}
\seealso{
\code{\link[base]{list.files}}
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
