% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parseInParanoid}
\alias{parseInParanoid}
\title{Parse inParanoid Table}
\usage{
parseInParanoid(MetaDataInParanoid)
}
\arguments{
\item{MetaDataInParanoid}{Provide a data frame where each row provides the meta data for a single inParanoid run. The resulting data frame must have 3 columns: 1) file path to table.inparanoid_output 2) Species A alias and 3) Species B alias}

\item{outDir}{Directory to write files}
}
\value{
The function will convert the table.inParanoid format to a two column edgelist and print the results to a text file SppA_SppB_orthologs.txt
}
\description{
Convert an inParanoid table format and output a table of ortholog edges
}
\examples{
load("Data/inParanoid_meta.rdata")
parseInParanoid(Meta_Data)
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
