% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_MultiSpp}
\alias{plot_MultiSpp}
\title{Plot MultiSpp co-appearance matrix}
\usage{
plot_MultiSpp(GeneMeta,order,CA_keep,sb=12, remove_0=TRUE,
    text_rotate=NULL, lwd=0.5, cex=0.2,
    my_palette=colorRampPalette(c("white", "lightyellow", "red","black"))(n = 100))
}
\arguments{
\item{GeneMeta}{Data frame that contains the project metadata. See \link{createGeneMeta}}

\item{order}{The order in which genes should be plotted.}

\item{CA_keep}{Sparse matrix of louvain community assignments generated from many runs of \code{\link{louvain}}}

\item{sb}{An integer specifying the increment of the sequence to plot. Example plots every 12th gene in the order.}

\item{remove_0}{A logical value indicating if the unclusterable genes should be plotted.}

\item{text_rotate}{Angle of module labels.}

\item{lwd}{Line weight}

\item{cex}{Font proportion}

\item{my_palette}{Specify your own color ramp for the heatmap}
}
\description{
Plot co-appearance heatmap of co-expression network generated across multiple species
}
\examples{
data("GeneMeta")
data("occurance")
data("MultiSppTrees")
plot_MultiSpp(GeneMeta = GeneMeta, order = MultiSpp_trees$order,
      sb=12, CA_keep = occurance, remove_0 = F, lwd = 1, cex = 0.5)
}
\seealso{
\code{\link[igraph]{cluster_louvain}}
}
\author{
Matthew Zinkgraf, \email{Matthew.Zinkgraf@wwu.edu}
}
