% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{checkGeneMetaFormat}
\alias{checkGeneMetaFormat}
\title{Check GeneMeta format}
\usage{
checkGeneMetaFormat(GeneMeta)
}
\arguments{
\item{GeneMeta}{Data frame that contains the project metadata. See \link{createGeneMeta}}
}
\value{
print error messages or PASS with summary of data
}
\description{
Check to make sure the GeneMeta object is formatted correctly and return summary of object
}
\examples{
data("eugra_potri_sapur_rpkm")
rpkm_var <- filterVariance(rpkm, variance = 0.1)
GeneMeta <- createGeneMeta(rpkm_var)
checkGeneMetaFormat(GeneMeta)
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
