% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createGeneMeta}
\alias{createGeneMeta}
\title{Create gene metadata object}
\usage{
createGeneMeta(rpkm)
}
\arguments{
\item{rpkm}{List object containing the rpkm values for each species.}
}
\description{
This function generates a data frame that contains the species and gene information, and columns to store module results.
}
\examples{
data("eugra_potri_sapur_rpkm")
rpkm_var <- filterVariance(rpkm, variance = 0.1)
GeneMeta <- createGeneMeta(rpkm_var)
}
\seealso{
\code{\link{checkGeneMetaFormat}}
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
