% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{louvain}
\alias{louvain}
\title{Louvain community detection}
\usage{
louvain(GeneMeta, edgelist, nruns, nThreads = 1)
}
\arguments{
\item{GeneMeta}{Data frame that contains the project metadata. See \link{createGeneMeta}}

\item{edgelist}{A data frame that contains three columns. The first two columns define the edges between gene_A and gene_B. Gene names must be converted to integer values. The third column defines the numeric edge weight.}

\item{nruns}{An interger defining the number of runs.}

\item{nThreads}{Integer specifying number of cores to be used by doParallel. Default = 1}
}
\description{
Run the Louvain clustering method multiple times on a user defined set of edges. For each run, the edgelist is shuffled so the louvain method begins at a random starting point in the network.
}
\examples{
data("combined_out")
data("GeneMeta")
nRuns = 100
results <- louvain(GeneMeta, edgelist = combined_out, nruns= nRuns)
}
\references{
Vincent D. Blondel, Jean-Loup Guillaume, Renaud Lambiotte, Etienne Lefebvre. 2008. Fast unfolding of communities in large networks. J. Stat. Mech. P10008
}
\seealso{
\code{\link[igraph]{cluster_louvain}}
}
\author{
Matthew Zinkgraf, \email{Matthew.Zinkgraf@wwu.edu}
}
\keyword{louvain}
