% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{weighted2rankList}
\alias{weighted2rankList}
\title{Correlation matrix to list of maximum}
\usage{
weighted2rankList(m, top=5, parallel_apply=FALSE, nThreads=2)
}
\arguments{
\item{m}{A maxtrix of correlation values}

\item{top}{An integer specifying the number values to return. Default = 5}

\item{nThreads}{Integer specifying number of cores to be used by parRapply. Default = 2}

\item{parallel_apply}{Logical value indicating if parRapply from parallel should be used in calulation Default = FALSE}
}
\description{
For each row in a matrix determine index values of the top (default=5) most correlated genes. An internal function used by getEdgelist()
}
\seealso{
\code{\link{getEdgelist}}
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
